added check if container exists

+++ start-db.sh
@@ -22,6 +22,13 @@ VOLUME_CONTAINER_PATH="/data/db"
 
 source setup.sh
 
+if [ "$(docker ps -q -f name=$CONTAINER_NAME)" ]; then
+    echo "A container with name $CONTAINER_NAME already exists."
+    echo "The container will be removed when stopped."
+    echo "To stop the container, run docker kill $CONTAINER_NAME"
+    exit 1
+fi
+
 docker run \
 --rm  -d --name=$CONTAINER_NAME \
 -e MONGODB_INITDB_ROOT_USERNAME=$ROOT_USER \
